Правка автозагрузки
