+++ web/index.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once("autoload.php"); // автозагрузка классов
+require_once('web_autoload.php'); // автозагрузка классов
 
 $localConfig = require(__DIR__ . '/../application/config/web-local.php');
 $config = ItForFree\rusphp\PHP\ArrayLib\Merger::mergeRecursivelyWithReplace(
